fix: support for old "codeLensProvider: true" server capability.
