+++ src/main/java/com/redhat/devtools/lsp4ij/features/diagnostics/LSPDiagnosticHandler.java
@@ -40,7 +40,7 @@ public class LSPDiagnosticHandler implements Consumer<PublishDiagnosticsParams> 
     @Override
     public void accept(PublishDiagnosticsParams params) {
         Project project = languageServerWrapper.getProject();
-        if (project == null || project.isDisposed()) {
+        if (project.isDisposed()) {
             return;
         }
         updateDiagnostics(params, project);
